requirements.txt
tabulate=0.9.0
keyboard=0.13.5
python-nmap=0.7.1